+++ test/controllers/cms/content_types_controller_test.rb
@@ -1,7 +1,0 @@
-require "test_helper"
-
-describe ContentTypesController do
-  # it "must be a real test" do
-  #   flunk "Need real tests"
-  # end
-end